+++ src/android/MCSdkListener.java
@@ -1,0 +1,25 @@
+package com.salesforce.marketingcloud.cordova;
+
+import android.app.PendingIntent;
+import android.content.Context;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.util.Log;
+
+import com.salesforce.marketingcloud.UrlHandler;
+
+public enum MCSdkListener implements UrlHandler {
+    INSTANCE;
+
+    public UrlHandler urlHandler = null;
+
+    @Nullable
+    @Override
+    public PendingIntent handleUrl(@NonNull Context context, @NonNull String s, @NonNull String s1) {
+        if (urlHandler != null) {
+            urlHandler.handleUrl(context, s, s1);
+        }
+
+        return null;  // The url will need to be handled by the Cordova dev in JS.
+    }
+}